Add more combat logic
Add logic for shooting guns
Add gun can be shot in 4 cardinal directions
Add Enemies to the game.
Rename FAL to AK47

+++ src/TresgalloP-GameProgramming2Final/Weapons/Weapon.cs
@@ -7,9 +7,9 @@ namespace TresgalloP_GameProgramming2Final.Weapons
     public abstract class Weapon
     {
         public string name;
-        public int damage;
-        public int accuracy;
-        public int magCapacity;
-        public int magCount;
+        public ushort damage;
+        public ushort accuracy;
+        public ushort magCapacity;
+        public ushort magCount;
     }
 }
